refacto: moved functions from main to tester.cpp
