+++ irc-tester/includes/testers.hpp
@@ -1,4 +1,14 @@
-//HEADER
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   testers.hpp                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: csweetin <csweetin@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/03/24 15:42:36 by csweetin          #+#    #+#             */
+/*   Updated: 2025/03/24 16:25:27 by csweetin         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #ifndef TESTERS_HPP
 #define TESTERS_HPP
@@ -8,9 +18,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int testRegistration(void);
-void compare(std::ifstream &output, std::ifstream &expectedOutput);
-int launchScript(const char *file);
-int openFile(std::ifstream &expectedOutput, const char *file);
+int test(const char *script, const char *expectedOutputFile,
+					 const char *testName);
 
 #endif
