--- irc-tester/src/testRegistration.cpp
@@ -1,7 +1,0 @@
-//HEADER
-#include "testers.hpp"
-
-int    testRegistration(void) {
-   
-   
-}
