+++ irc-tester/tester-scripts/test-nick.sh
